Fix set encoding
Before, there were three ways to set a file or string's character encoding:
explicitly state the encoding's name in the constructor or set method, get
the library to detect the encoding with the special 'auto' keyword, and
use the value returned by mb_detect_encoding() by using a null value. Eh,
I didn't like how that complicated things. It was dirty.

Now, you either explicitly state the encoding in the constructor or it is
implicitly set on construction to the value of mb_detect_encoding(). Of
course, you can over-ride the explicit or implicit value with the
setEncoding() method whenever you want.